Reorganize StaticString implementation
