+++ crates/cgp-field/src/traits/mod.rs
@@ -1,6 +1,5 @@
 mod build_field;
 mod extract_field;
-mod format;
 mod from_fields;
 mod from_variant;
 mod has_builder;
@@ -14,6 +13,8 @@ mod map_type_ref;
 mod match_str;
 mod nat;
 mod partial_data;
+mod static_format;
+mod static_string;
 mod take_field;
 mod to_fields;
 mod transform_map;
@@ -21,7 +22,6 @@ mod update_field;
 
 pub use build_field::*;
 pub use extract_field::*;
-pub use format::*;
 pub use from_fields::*;
 pub use from_variant::*;
 pub use has_builder::*;
@@ -35,6 +35,8 @@ pub use map_type_ref::*;
 pub use match_str::*;
 pub use nat::*;
 pub use partial_data::*;
+pub use static_format::*;
+pub use static_string::*;
 pub use take_field::*;
 pub use to_fields::*;
 pub use transform_map::*;
