+++ crates/cgp-field/src/types/char.rs
@@ -1,7 +1,7 @@
 use core::fmt::Display;
 use core::marker::PhantomData;
 
-use crate::traits::{MaybeChars, StaticFormat};
+use crate::traits::StaticFormat;
 
 /**
     The `Char` type, a.k.a. `ι`, is used to represent _type-level_ list of
@@ -59,25 +59,4 @@ where
     fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         <Self as StaticFormat>::fmt(f)
     }
-}
-
-impl<const CHAR: char, Tail> StaticFormat for Char<CHAR, Tail>
-where
-    Tail: StaticFormat,
-{
-    fn fmt(f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-        write!(f, "{CHAR}")?;
-        Tail::fmt(f)
-    }
-}
-
-impl<const CHAR: char, Tail> MaybeChars for Char<CHAR, Tail>
-where
-    Tail: MaybeChars,
-{
-    const LEN: usize = Tail::LEN + 1;
-
-    const VALUE: Option<char> = Some(CHAR);
-
-    type Next = Tail;
 }
